// src/types/structures/message.ts
export interface Embed {
  /** Title of embed */
  title?: string;
  /** Type of embed (always "rich" for webhook embeds) */
  type?: string;
  /** Description of embed */
  description?: string;
  /** URL of embed */
  url?: string;
  /** Timestamp of embed content */
  timestamp?: string;
  /** Color code of the embed */
  color?: number;
  /** Footer information */
  footer?: EmbedFooter;
  /** Image information */
  image?: EmbedImage;
  /** Thumbnail information */
  thumbnail?: EmbedThumbnail;
  /** Video information */
  video?: EmbedVideo;
  /** Provider information */
  provider?: EmbedProvider;
  /** Author information */
  author?: EmbedAuthor;
  /** Fields information */
  fields?: EmbedField[];
}

/**
 * Embed footer object
 */
export interface EmbedFooter {
  /** Footer text */
  text: string;
  /** URL of footer icon (only supports http(s) and attachments) */
  icon_url?: string;
  /** A proxied url of footer icon */
  proxy_icon_url?: string;
}

/**
 * Embed image object
 */
export interface EmbedImage {
  /** Source url of image (only supports http(s) and attachments) */
  url: string;
  /** A proxied url of the image */
  proxy_url?: string;
  /** Height of image */
  height?: number;
  /** Width of image */
  width?: number;
}

/**
 * Embed thumbnail object
 */
export interface EmbedThumbnail {
  /** Source url of thumbnail (only supports http(s) and attachments) */
  url: string;
  /** A proxied url of the thumbnail */
  proxy_url?: string;
  /** Height of thumbnail */
  height?: number;
  /** Width of thumbnail */
  width?: number;
}

/**
 * Embed video object
 */
export interface EmbedVideo {
  /** Source url of video */
  url?: string;
  /** A proxied url of the video */
  proxy_url?: string;
  /** Height of video */
  height?: number;
  /** Width of video */
  width?: number;
}

/**
 * Embed provider object
 */
export interface EmbedProvider {
  /** Name of provider */
  name?: string;
  /** URL of provider */
  url?: string;
}

/**
 * Embed author object
 */
export interface EmbedAuthor {
  /** Name of author */
  name: string;
  /** URL of author (only supports http(s)) */
  url?: string;
  /** URL of author icon (only supports http(s) and attachments) */
  icon_url?: string;
  /** A proxied url of author icon */
  proxy_icon_url?: string;
}

/**
 * Embed field object
 */
export interface EmbedField {
  /** Name of the field */
  name: string;
  /** Value of the field */
  value: string;
  /** Whether or not this field should display inline */
  inline?: boolean;
}

/**
 * Reaction object
 */
export interface Reaction {
  /** Times this emoji has been used to react */
  count: number;
  /** Whether the current user reacted using this emoji */
  me: boolean;
  /** Emoji information */
  emoji: PartialEmoji;
}

/**
 * Partial emoji object
 */
export interface PartialEmoji {
  /** Emoji ID */
  id?: string | null;
  /** Emoji name */
  name?: string | null;
  /** Whether this emoji is animated */
  animated?: boolean;
}
// ...
export interface MessageComponent {
  /** Component type */
  type: number;
  /** Developer-defined identifier for the component */
  custom_id?: string;
  /** Whether the component is disabled */
  disabled?: boolean;
  /** One of button styles */
  style?: number;
  /** Text that appears on the button */
  label?: string;
  /** Name, id, and animated */
  emoji?: PartialEmoji;
  /** URL for link-style buttons */
  url?: string;
  /** Options for select menus */
  options?: SelectOption[];
  /** List of channel types to include in the channel select component */
  channel_types?: number[];
  /** Placeholder text if nothing is selected */
  placeholder?: string;
  /** Minimum number of items that must be chosen */
  min_values?: number;
  /** Maximum number of items that can be chosen */
  max_values?: number;
  /** A list of child components */
  components?: MessageComponent[];
}

/**
 * Select option object
 */
export interface SelectOption {
  /** User-facing name of the option */
  label: string;
  /** Dev-defined value of the option */
  value: string;
  /** Additional description of the option */
  description?: string;
  /** ID, name, and animated */
  emoji?: PartialEmoji;
  /** Will show this option as selected by default */
  default?: boolean;
}